requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Deprecated==1.2.13
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.0
fastapi==0.70.0
fastapi-mail==1.0.2
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
jose==1.0.0
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
redis==4.0.2
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.28
starlette==0.16.0
toml==0.10.2
typing-extensions==4.0.1
uvicorn==0.16.0
wrapt==1.13.3
deta
